\name{ltsk-interval}
\alias{ltsk-interval}
\alias{cal.gamma}
\alias{check_input}
\alias{check_na}
\alias{dadjustsills}
\alias{dfitvariogram}
\alias{dnb.index}
\alias{dsample.pps}
\alias{dsample.strata}
\alias{dsmooth.variogram}
\alias{dsubsample2}
\alias{dvariogram}
\alias{firstpeak}
\alias{tritomat}
\alias{vexpn}
\alias{vgaun}
\alias{vmten}
\alias{vsphn}
\alias{work.calgamma}
\alias{work.kriging}
\alias{working.compvariogmodels1}
\alias{working.fitvariog1}
\alias{working.smoothvariogram}
\alias{working.dnb}
\alias{working.ltsk}
\alias{working.variog}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions to ltsk
}
\description{
These functions are working R functions that are called by the ltsk function. They should not be directly used.
}

