\name{ltsk-package}
\alias{ltsk-package}
\docType{package}
\title{
Local time space kriging
}
\description{
ltsk is a collection of programs for local Kriging with large spatiotemporal data sets.
The major methods implmented include product-sum model based spatiotemporal variogram
estimation and Kriging within a local neighborhood. An important feature is that ltsk
enable working with large data through sub-sampling of neighbor-hood in spatiotemporal
balanced way.
}
\details{
\tabular{ll}{
Package: \tab ltsk\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-29\cr
License: \tab GPL2.0\cr
}
}
\author{
Jun chen (wdidwlia@gmail.com)
Jin Chen (jc.chenjin@gmail.com)
Dong Liang (dong-liang@uiowa.edu)
Naresh Kumar (NKumar@med.miami.edu)
}
\references{
Iaco, S. De & Myers, D. E. & Posa, D., 2001. "Space-time analysis using a general product-sum model," Statistics & Probability Letters, Elsevier, vol. 52(1), pages 21-28, March.
}
\keyword{package}

