\name{ltsk-interval}
\alias{ltsk-interval}
\alias{dadjustsills}
\alias{firstpeak}
\alias{vmte}
\alias{vpown}
\alias{check_input}
\alias{vopw}
\alias{dsphn}
\alias{dfitvariogram}
\alias{vsph}
\alias{work.calgamma}
\alias{tritomat}
\alias{dsample.pps}
\alias{vexpn}
\alias{working.compvariogmodels1}
\alias{dsample.strata}
\alias{vexp}
\alias{working.fitvariog1}
\alias{dsmooth.variogram}
\alias{working.ltsk}
\alias{working.ltsk.par}
\alias{dsubsample}
\alias{vgau}
\alias{working.smoothvariogram}
\alias{dvariogram}
\alias{vmten}
\alias{work.kriging}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions to ltsk
}
\description{
These functions are working R functions that are called by the ltsk function. They should not be directly used.
}

